Move openapi to visibility directory.
