+++ apis/visibility/openapi/doc.go
@@ -15,6 +15,6 @@ limitations under the License.
 */
 
 // To avoid the following error from controller-gen, we need to have this file.
-// "-: build constraints exclude all Go files in /home/prow/go/src/sigs.k8s.io/kueue/apis/visibility/v1alpha1/openapi"
+// "-: build constraints exclude all Go files in /home/prow/go/src/sigs.k8s.io/kueue/apis/visibility/openapi"
 
 package openapi
